refactor(cli): split CLI TUI god files (cliState, App child-control toggle, ChildControlPicker scroll state)
- Split cliState.ts (532 lines, ~25-field StreamSlice store) into a
  state/cliState/ directory of cohesive slice modules (session, streams,
  parent-stream, focus, foreground overlays, exit hint, codex preference,
  cross-cutting removeStream/reset) behind an index.ts barrel that preserves
  the existing `cliState` object and export surface, so no consumer imports
  changed.
- Moved App.tsx's local childControlMode/childControlEscapeAction useState
  into signal state (new foregroundOverlaySlice), per the CLAUDE.md rule that
  view-level toggles belong in shared signal state, not per-component state.
- Extracted ChildControlPicker.tsx's bespoke TaskDetailScrollState machine
  into state/taskDetailScroll.ts, alongside the useScrollableOffset.ts and
  transcriptScroll.ts modules it was paralleling, shrinking the component
  from 1198 to 911 lines.
